# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API GPE",
    "version": "1.0.0",
    "description": "Documentação da API do sistema GPE desenvolvida em Node.js com Express e MySQL."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/cadastro": {
      "post": {
        "tags": ["Autenticação","Gestores"],
        "summary": "Cadastra um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "João Silva"
                  },
                  "email": {
                    "type": "string",
                    "example": "joao@email.com"
                  },
                  "senha": {
                    "type": "string",
                    "example": "123456"
                  },
                  "tipo": {
                    "type": "string",
                    "enum": ["aluno", "pai", "professor", "gestor"],
                    "example": "aluno"
                  },
                  "telefone": {
                    "type": "string",
                    "example": "(11) 99999-9999"
                  },
                  "disciplina": {
                    "type": "string",
                    "example": "Matemática"
                  },
                  "turma_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "pai_id": {
                    "type": "integer",
                    "example": 2
                  }
                },
                "required": ["nome", "email", "senha", "tipo"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cadastro realizado com sucesso"
          },
          "400": {
            "description": "Erro de validação"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
...
    "/api/alunos": {
      "get": {
        "tags": ["Alunos"],
        "summary": "Lista todos os alunos",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "500": {
            "description": "Erro ao listar alunos"
          }
        }
      }
    },
    "/api/alunos/{id}/turma": {
      "patch": {
        "tags": ["Alunos"],
        "summary": "Atualiza a turma de um aluno",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "turma_id": {
                    "type": "integer",
                    "example": 2
                  }
                },
                "required": ["turma_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Turma do aluno atualizada com sucesso"
          },
          "404": {
            "description": "Aluno não encontrado"
          },
          "500": {
            "description": "Erro ao atualizar turma do aluno"
          }
        }
      }
    },
    "/api/pais": {
      "get": {
        "tags": ["Pais"],
        "summary": "Lista todos os pais",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "500": {
            "description": "Erro ao listar pais"
          }
        }
      }
    },
    "/api/professores": {
      "get": {
        "tags": ["Professores","Gestores"],
        "summary": "Lista todos os professores",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "500": {
            "description": "Erro ao listar professores"
          }
        }
      }
    },
    "/api/usuarios": {
      "get": {
        "tags": ["Gestores"],
        "summary": "Lista todos os usuários",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "500": {
            "description": "Erro ao listar usuários"
          }
        }
      }
    },
...
  }
}